Added New User Integration and Documentation
Main focus of this commit was adding the functionality of taking the data from the New User window and creating the user in the database.  Checks are in place to catch if someone is trying to create a user with a name that is already in the database allowing a message to pop up explaining to the user what is going on.  Additional documentation and type hinting has been added to add clarity to the program.