Visits Table Created & Added Functionality to New User Controller
The Visits table along with associated classes, methods, and tests have been created.  Additionally functionality has been extended for the New User Controller to include both the creation of a new user and the creation of a new visit with the new user's id, as well as closing the new user window at the end.