Added Functionality for Logging In and Refactor UI to One Window
Several changes in this commit.  Logging in from the UI was implemented along with a check that ensures users can't log in more than once.  Along the way, incorporating a scrollbar onto the members listbox resulted in a refactoring of the user interface to be based around frames, allowing the window to change rather than creating additional popup windows.