Visit Sign Out Completed
After biting off more than was realistic for a single commit in regards to the sign out functionality, I took several steps back and instead focused on setting the sign out time for the visit, thus removing the record from the signed in list.  After consulting Nicholas, I'm going to restructure some of the data on the logged out aspect of the database.