Sign Out Functionality Added
This commit includes most of the functionality needed to add to the database; however, it is still quite buggy.  When all of the information is added properly, the application just fine.  However, the unique qualifier on the project name is causing some problems and the errors aren't yet handled properly.  I think that my removing this and returning the primary key instead of trying to use unique name should solve this problem.