Refactored the Code to Remove Load Based on Name
Refactored code so the "load" method was renamed to be "load_by_name" and a new method was created using the user_id.  Then, we tracked the usage of "load_by_name" and refactored the code to use the new "load" method instead.  This also involved updating the method in the visit class "get_logged_in_users" to return both names and user_ids.  There remains only one use of "load_by_id" which is as a check in the new_user_controller to determine if a user with the given name already exists prior to committing the new user to the database.  This final usage may remain unchanged for the time being.