Project Load Method Updated to Primary Key
The load functionality of the project class has been altered to use the primary key instead of the project name for searching.  As a result, the need for projects to have unique names was also eliminated and the initialize_database project was updated accordingly.  The database tables for users, visits, projects, usage_log, materials_consumed, and visits_projects were all dropped and recreated to avoid conflicting data in the database.  This was needed for the project_name unique keyword to be removed.