#3 Removal of Final Load_Last Methods
The final methods that utilized the load_last logic have been removed and replaced by a standard load method utilizing the primary key. This should be the end of the refactoring and will allow work to begin on Issue #4 to add functionally which permits multiple visits to be attributed to a single project.