refactor: Move DB resources from domain pkg to flask
All of the database calls were under the domain package with the domain
object definitions. The UI / Client depend on the domain modules, but
not on the DB. To avoid adding unused database dependencies to the
client, all of the database access logic was moved to the
ws_login_flaskr.repositories package.