feat! Create script for populating test users
Created script for populating the system with a bunch of test users.
Populating the db with this number of users will allow monitoring of the
perforance of the application when it starts getting full.

The date format passed back from the api was discovered to be
problimatic to parse. The client may have been using a string instead of
actually converting the value into a date time.  As a result, the api is
starting to return dates back in isoformat.  This has been done only on
the users-details api thus far.