feat: Add flask-login and basic permissions to project
This is part of an effort to better lock down the access to the data,
and prevent just arbitrary access to all data for all users.

A user of the system must now declare who they are, although no
authentication is in place in order to prove that who they say they are
is true.  Once there user is declared, permissions will be locked down
so they can only see their visits and their projects.

Some permissioning work has been added as part of /api/visits endpoint,
but this is more proof of concept than complete work.

All endpoints should still accept the access-token cookie as an
alternative to this session login.